Move publish endpoint to the revision controller, remove parent arg
